requirements.txt
requests
python-dateutil
python-dotenv
pandas
openpyxl
gspread
google-auth 
google-auth-oauthlib 
google-auth-httplib2
google-api-python-client